S13P31A703 fix:pace int 변경

+++ backend/runnity/src/main/java/com/runnity/history/domain/RunLap.java
@@ -27,7 +27,7 @@ public class RunLap extends BaseEntity {
     private Integer durationSec;
 
     @Column(nullable = false)
-    private Float pace;
+    private Integer pace;
 
     @Column(nullable = false)
     private Integer bpm;
